requirements.txt
Django==1.11
pytz==2017.2
djangorestframework
pandas
django-cors-headers
scipy
sklearn
pydot